+++ ui/src/main/java/ru/tinkoff/acquiring/sdk/redesign/tpay/nav/TpayNavigation.kt
@@ -1,0 +1,20 @@
+package ru.tinkoff.acquiring.sdk.redesign.tpay.nav
+
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.receiveAsFlow
+import ru.tinkoff.acquiring.sdk.redesign.tpay.Tpay
+
+internal class TpayNavigation() {
+    private val events = Channel<Event>()
+    val flow = events.receiveAsFlow()
+
+    suspend fun send(event: Event) {
+        events.send(event)
+    }
+
+    sealed interface Event {
+
+        class GoToTinkoff(val deeplink: String) : Event
+        class Close(val result: Tpay.Result) : Event
+    }
+}
